--- CAN SEND/CAN SEND/Debug/CAN SEND.hex
@@ -1,83 +1,0 @@
-:100000000C9434000C9452010C9446000C9446005D
-:100010000C9446000C9446000C9446000C94460048
-:100020000C9446000C9446000C9446000C94460038
-:100030000C9446000C9446000C9446000C94460028
-:100040000C9446000C9446000C9446000C94460018
-:100050000C9446000C9446000C9446000C94460008
-:100060000C9446000C94460011241FBECFEFD8E03C
-:10007000DEBFCDBF21E0A0E0B1E001C01D92A23003
-:10008000B207E1F70E9458020C9487020C9400001A
-:100090008CE284B980E58CBD81E08DBD08958EBD74
-:1000A0000DB407FEFDCF8EB50895811103C02A9AC5
-:1000B000229A08952A98229A0895CF93C82F81E012
-:1000C0000E94550083E00E944F008C2F0E944F0039
-:1000D00080E00E944F00C82F80E00E9455008C2FC6
-:1000E000CF910895CF93DF93D82FC62F81E00E9440
-:1000F000550082E00E944F008D2F0E944F008C2FF0
-:100100000E944F0080E00E945500DF91CF9108953A
-:100110001F93CF93DF93182FD62FC42F81E00E9417
-:10012000550085E00E944F00812F0E944F008D2FC7
-:100130000E944F008C2F0E944F0080E00E945500CB
-:10014000DF91CF911F91089581E00E94550080ECCE
-:100150000E944F0080E00E94550008950F931F9366
-:10016000CF93DF93EC018C011695079516950795B3
-:1001700016950795602F80E20E947200C295CC0F01
-:10018000C07E6C2F81E20E947200602F84E20E9488
-:1001900072006C2F85E20E947200DF91CF911F9157
-:1001A0000F9108950F931F93CF93DF93EC018C0170
-:1001B000169507951695079516950795602F80E07B
-:1001C0000E947200C295CC0FC07E6C2F81E00E940D
-:1001D0007200602F84E00E9472006C2F85E00E9404
-:1001E0007200602F88E00E9472006C2F89E00E94EC
-:1001F0007200602F80E10E9472006C2F81E10E94EA
-:100200007200602F84E10E9472006C2F85E10E94D1
-:100210007200602F88E10E9472006C2F89E10E94B9
-:100220007200DF91CF911F910F9108951F93CF938B
-:10023000DF93182FD62FC42F8EE00E945D00829589
-:1002400086958770843079F4612F8AE20E9472006B
-:100250006D2F89E20E9472004C2F67E088E20E94B5
-:10026000880081E001C080E0DF91CF911F91089567
-:10027000E9E6F0E080818260808380818E7F8083E8
-:10028000E89A089541E061E08BE20E94880040E036
-:1002900062E08CE20E94880040E061E08CE20E9413
-:1002A000880008951F920F920FB60F9211242F937A
-:1002B0003F934F935F936F937F938F939F93AF93EE
-:1002C000BF93EF93FF9340E062E08CE20E948800CE
-:1002D00040E061E08CE20E94880081E090E0909331
-:1002E000010180930001FF91EF91BF91AF919F9128
-:1002F0008F917F916F915F914F913F912F910F90CF
-:100300000FBE0F901F9018950E9448000E94380160
-:100310000E94A40040E860EE8FE00E9488008EE01A
-:100320000E945D008295869587708430C1F741E018
-:1003300061E981E00E94160180E090E00E94AE0039
-:1003400080E090E00E94D20044E064E080E60E94F9
-:1003500088000E94420140E060EE8FE00E94880029
-:100360008EE00E945D00829586958770C9F708959A
-:100370008F929F92AF92BF92EF92FF920F931F9333
-:10038000CF93C82F4A015B01790181E00E9455009B
-:1003900082E00E944F008C2F8295807F805D0E94BA
-:1003A0004F00802F82958695869583700E944F001E
-:1003B00004FF3AC0D501C401689412F8B695A79518
-:1003C000979587951694D1F70E944F00282D229576
-:1003D000220F207ED501C401072E7BE1B695A7959B
-:1003E000979587957A95D1F7702D822B88600E941A
-:1003F0004F00D501C401072E73E1B695A7959795D7
-:1004000087957A95D1F7702D0E944F00D501C401D0
-:10041000072E7BE0B695A795979587957A95D1F7A6
-:10042000702D0E944F0012C0D501C401689412F8CB
-:10043000B695A795979587951694D1F70E944F008A
-:10044000882D8295880F807E0E944F00802F8F70AC
-:1004500005FD02C090E001C090E4892B0E944F008E
-:100460000F7010E0101611066CF4C0E0F701EC0FED
-:10047000F11D80810E944F00CF5F8C2F90E080178C
-:100480009107A4F380E00E94550081E00E9455008E
-:1004900081E80E944F0080E00E945500CF911F919B
-:1004A0000F91FF90EF90BF90AF909F908F90089525
-:1004B000CF93DF93CDB7DEB728970FB6F894DEBFA2
-:1004C0000FBECDBF0E9484017894FE01319688E072
-:1004D000DF011D928A95E9F781E0898382E08A83B2
-:1004E00083E08B8303E09E012F5F3F4F41EE57E097
-:1004F00060E070E080E00E94B801BFEF23ED80E390
-:10050000B15020408040E1F700C00000EBCFF894EC
-:02051000FFCF1B
-:00000001FF
