+++ CAN_SEND/CAN SEND/Debug/makedep.mk
@@ -1,0 +1,6 @@
+################################################################################
+# Automatically-generated file. Do not edit or delete the file
+################################################################################
+
+main.c
+
